+++ spec/dummy/app/controllers/admin_controller.rb
@@ -1,4 +1,4 @@
-# FIXME: mover a pg_engine
+# FIXME!: mover a pg_engine
 class AdminController < ApplicationController
   include PgEngine::RequireSignIn